llm-evaluation-tool/requirements.txt
flask==3.0.0
flask-cors==4.0.0
gspread==6.0.0
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
openai==1.12.0
anthropic==0.18.1
google-generativeai==0.3.2
python-dotenv==1.0.1